move condition response getter to new class
