+++ condition-checker/src/main/java/com/mersey/rowing/club/condition_checker/controller/config/ClockConfig.java
@@ -1,16 +1,15 @@
 package com.mersey.rowing.club.condition_checker.controller.config;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
 import java.time.Clock;
 import java.time.ZoneId;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class ClockConfig {
 
-    @Bean
-    public Clock clock() {
-        return Clock.system(ZoneId.of("Europe/London"));
-    }
+  @Bean
+  public Clock clock() {
+    return Clock.system(ZoneId.of("Europe/London"));
+  }
 }
